extract/requirements.txt
bson==0.5.10
DateTime==5.2
dnspython==2.4.2
greenlet==2.0.2
logging==0.4.9.6
numpy==1.25.2
pandas==2.0.3
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
zope.interface==6.0
